separate classes in views
