+++ registration/views/__init__.py
@@ -1,0 +1,3 @@
+from email import *
+from register import *
+from confirm import *
